# openapi.yaml
openapi: 3.0.3
info:
  title: Dropbox Media Server API
  description: API for shared media gallery, public file requests, and administration.
  version: 1.6.0
servers:
  - url: /
paths:
...
  /api/droppr/video-meta:
    get:
      summary: Get video metadata by filesystem path
      security:
        - BearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Video metadata
  /api/droppr/preview:
    get:
      summary: Get file preview by filesystem path
      security:
        - BearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Image data

  # File Requests
...
components:
...
  securitySchemes:
    XAuth:
      type: apiKey
      in: header
      name: X-Auth
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...